app/requirements.txt
python-dotenv
openai
sentence-transformers
scikit-learn
numpy
pydantic